Separate bowling_test.rb for each class
